feat(bom): Allow decimal values for total_required in BOM items
- Changed validation rule for total_required from 'integer' to 'numeric' in BomItemController for both store and update methods to support decimal quantities
- Removed integer casting for total_required in BomItem model to allow decimal storage
- Updated machine types validation in MachineController to include new values (LASPEN, LAS_MIG, PHOSPATHING, PACKING) and switched to Validator facade for better error handling in store and update methods
- Modified database migration to adjust total_required column type to decimal (truncated in diff)

+++ database/migrations/2025_12_18_183314_create_bom_items_table.php
@@ -27,7 +27,7 @@ return new class extends Migration
                 ->onDelete('restrict');
 
             $table->decimal('quantity_per_unit');
-            $table->integer('total_required');
+            $table->decimal('total_required');
 
             $table->integer('allocated');
             $table->integer('realized');
